Build is running with Andy's new stuff. Very good time for a commit. Right now, only RestAdapter is supported.  Need to finish the others.
simulator/pom.xml: the JBoss repository moved.  Needed to update for the JBoss dependencies
FullResponseSoapAdapter: for whatever reason, Definition.getAllBindings was not found in the method declaration.  This is for axis-wsdl4j.
RestAdapter: uses Jackson to prepare BSF global objects as NativeObject rather than dynamic classes.  To move this into the other Adapters
PojoClassGenerator: Added an AtomicLong rather than timestamp.  This class will be deleted when other BaseAdapters are fixed
 ScriptExecutionService: Got two almost identical methods here.  Not sure I like signatures on either of them.  Its a todo
 RestTransportationIntegrationTest: added a couple asserts for debugging
 ScriptExecutionServiceTest: added some new tests.  They need cleanup.  They cover using the NativeObjects for BSF.
 simulator-dependencies-tomcat: updated to javax.activation version due to a Maven central issue
 test-system module: added a load test.