General comment:  There's way too much CCN on the class level due to layers of inheritance.  Overall, the inheritance structure needs to be broken up and switched to interfaces with delegates.  Its very very hard to debug right now.
ScenarioExecutionWrapperTest: Ignores on the Ruby tests.  Currently not supporting
XMLAdapter: now extends NativeObjectScriptingAdapter
BaseAdapter: changed the result StructuredPojo generator (getSimulatorPojo) to protected.  
CsvAdapter: overrides getSimulatorPojo so that it can override the root element with the CSV Content Param
JsonAdapter: overrides getSimulatorPojo so that it can override the root element with the Json Content Param
SoapAdapter: overrides getSimulatorPojo so that it can override the root element with the DEFAULT_PAYLOAD_KEY
SimulatorPojoPopulatorImpl: Provides support for overriding root element in map.